Group `ty::Predicate` internals in one file
